refactor: WhisperingButton
